% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-GEN.R
\name{abbreviations}
\alias{abbreviations}
\alias{abbrs}
\title{Commonly used abbreviations in \R code}
\usage{
abbreviations(expand = FALSE)

abbrs(expand = FALSE)
}
\arguments{
\item{expand}{Whether to expand the full expression column and return the data in long format. If \code{FALSE}, a "nested" list column \code{full_expressions} will be
returned, meaning the values in column \code{abbreviation} will be unique.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} listing an opinionated set of abbreviations commonly used when writing \R code. It can be used as a reference, for
example, to check availability when considering using some abbreviation in a function or argument name.
}
