% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\docType{data}
\name{data_msgs}
\alias{data_msgs}
\title{\R condition messages}
\format{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{id} and \code{value}.
}
\usage{
data_msgs
}
\description{
A \link[tibble:tbl_df-class]{tibble} with all \R condition messages included in this package, together with their \code{id}. The latter can be fed to \code{\link[=msg]{msg()}}.
}
\details{
Currently, \R condition messages with the following \code{id}s and \code{arguments} are available:\tabular{ll}{
   \strong{id} \tab \strong{arguments} \cr
   pkg_required \tab \code{pkg} \cr
}
}
\seealso{
R message functions
\code{\link{msg}()}
}
\concept{rmsg}
\keyword{datasets}
