% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{ls_file_snips}
\alias{ls_file_snips}
\title{List all available R Markdown file snippets}
\usage{
ls_file_snips()
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
Lists all of the R Markdown snippets shipped with this package, together with the paths where they're located on the filesystem.
}
\seealso{
R Markdown file snippet functions:
\code{\link[=snip_path]{snip_path()}}
}
\concept{rmdsnips}
