% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-GEN.R
\name{ls_rmd_snippets}
\alias{ls_rmd_snippets}
\title{List all available R Markdown file snippets}
\usage{
ls_rmd_snippets()
}
\value{
\code{pkgsnip::return_label("data")}
}
\description{
This function lists all of the R Markdown snippets shipped with this package, together with the paths where they're located on the filesystem.
}
\seealso{
R Markdown file snippet functions
\code{\link{rmd_snippet_path}()}
}
\concept{rmdsnips}
