% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{ls_snippets}
\alias{ls_snippets}
\title{List all snippets shipped with this package}
\usage{
ls_snippets()
}
\value{
A vector of \url{paths}.
}
\description{
List all snippets shipped with this package
}
