% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-GEN.R
\name{md_snippet}
\alias{md_snippet}
\title{Get predefined Markdown snippet}
\usage{
md_snippet(name = md_snippets()$name)
}
\arguments{
\item{name}{The snippet name.}
}
\value{
A character scalar.
}
\description{
Get predefined Markdown snippet
}
\seealso{
Markdown snippet functions
\code{\link{md_snippets}()}
}
\concept{mdsnips}
