% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{md_snippets}
\alias{md_snippets}
\title{Get a table of all Markdown snippets included in this package}
\usage{
md_snippets()
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
This simply returns a \link[tibble:tbl_df]{tibble} listing all Markdown snippets together with their \code{name} which can be provided as \code{\link[=md_snippet]{md_snippet()}}'s \code{name}
argument.
}
\details{
Currently, Markdown snippets with the following \code{names} are available:

\link{1} "\n\n| \strong{name}             |\n|:---------------------|\n| \code{rstudio_addin_hint} |\n\n"
}
\seealso{
roxygen2 label functions
\code{\link{param_label}()},
\code{\link{return_label}()},
\code{\link{roxygen_labels}()}
}
\concept{roxygen2label}
