% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-GEN.R
\name{md_snips}
\alias{md_snips}
\title{Get a table of all Markdown snippets included in this package}
\usage{
md_snips()
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
This simply returns a \link[tibble:tbl_df-class]{tibble} listing all Markdown snippets together with their \code{name} which can be provided as \code{\link[=md_snip]{md_snip()}}'s \code{name}
argument.
}
\details{
Currently, Markdown snippets with the following \code{names} are available:\tabular{l}{
   \strong{name} \cr
   \code{rstudio_addin_hint} \cr
}
}
\seealso{
Markdown snippet functions
\code{\link{md_snip}()}
}
\concept{mdsnips}
