% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{messages}
\alias{messages}
\title{Get a table of all available R condition messages}
\usage{
messages()
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
This simply returns a \link[tibble:tbl_df-class]{tibble} listing all R condition messages included in this package, together with their \code{name} which can be provided
as \code{\link[=msg]{msg()}}'s \code{name} argument.
}
\details{
Currently, R condition messages with the following \code{names} and \code{arguments} are available:\tabular{ll}{
   \strong{name} \tab \strong{arguments} \cr
   \code{pkg_required} \tab \code{pkg} \cr
   \code{pkg_2} \tab \code{pkg} \cr
}
}
\seealso{
R message functions
\code{\link{msg}()}
}
\concept{rmsg}
