% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{msg}
\alias{msg}
\title{Get predefined \R condition message}
\usage{
msg(id = data_msgs$id, ...)
}
\arguments{
\item{id}{The message identifier. One of \code{"pkg_required"}.}

\item{...}{Further named arguments used to tailor the message to your needs. Not all messages require additional arguments, see \code{\link{data_msgs}} for
an overview. If a required argument is not explicitly provided, it is searched for in the \link[=parent.frame]{parent frames}.}
}
\value{
A character scalar.
}
\description{
Get predefined \R condition message
}
\examples{
pkgsnip::msg(id = "pkg_required",
             pkg = "some_pkg")
}
\seealso{
R message functions
\code{\link{data_msgs}}
}
\concept{rmsg}
