% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{param_label}
\alias{param_label}
\title{Get predefined parameter label}
\usage{
param_label(name = roxygen_labels(type = "param")$name)
}
\arguments{
\item{name}{The label name. See \code{\link[=roxygen_labels]{roxygen_labels()}} for possible values.}
}
\value{
A character scalar.
}
\description{
These are pre-defined labels intended to be used to document parameters using \link[roxygen2:roxygen2-package]{roxygen2}'s
\href{https://roxygen2.r-lib.org/articles/rd.html#functions}{\verb{@param}} tag.
}
\details{
The parameters can be inserted using \href{https://roxygen2.r-lib.org/articles/rd-formatting.html#dynamic-r-code-1}{inline R code} as follows:

\verb{#' @param start_date `r pkgsnippets::param_label("start_date")` }
\verb{#' @param end_date `r pkgsnippets::param_label("end_date")` }
\verb{#' @param use_cache `r pkgsnippets::param_label("use_cache")` }
\verb{#' @param cache_lifespan `r pkgsnippets::param_label("cache_lifespan")` }

Note that the above only works in \href{https://www.tidyverse.org/blog/2020/03/roxygen2-7-1-0/}{roxygen2 7.1.0} and above.
}
\seealso{
roxygen2 label functions
\code{\link{return_label}()},
\code{\link{roxygen_labels}()}
}
\concept{roxygen2label}
