% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-package.R
\docType{package}
\name{pkgsnip-package}
\alias{pkgsnip}
\alias{pkgsnip-package}
\title{pkgsnip: Standardized roxygen2 Labels and (R) Markdown Snippets for Package Authors}
\description{
pkgsnip ships ready-to-go (R) Markdown snippets and labels intended to be used in roxygen2 documentation or as knitr child documents. Also included are some predefined R condition messages and a table of abbreviations commonly used in R code. All of them are designed to reduce unnecessary repetition and standardize R package documentation wherever possible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkgsnip.rpkg.dev}
  \item \url{https://gitlab.com/rpkg.dev/pkgsnip}
  \item Report bugs at \url{https://gitlab.com/rpkg.dev/pkgsnip/-/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim@rpkg.dev} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

Authors:
\itemize{
  \item Salim Brüggemann \email{salim@rpkg.dev} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})
}

}
\keyword{internal}
