% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-package.R
\docType{package}
\name{pkgsnippets-package}
\alias{pkgsnippets}
\alias{pkgsnippets-package}
\title{pkgsnippets: Provides Handy (R) Markdown Snippets for Package Authors}
\description{
pkgsnippets ships ready-to-go (R) Markdown snippets intended to be
    used as knitr child documents. Use `pkgsnippets::file()` to get the path of
    a snippet and feed it to knitr's `child` chunk option.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/salim_b/r/pkgs/pkgsnippets}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/pkgsnippets/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
