% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-GEN.R
\name{rmd_snip_path}
\alias{rmd_snip_path}
\title{Get R Markdown snippet file path}
\usage{
rmd_snip_path(name = ls_rmd_snips()$name)
}
\arguments{
\item{name}{The name of a snippet. Possible values include:
\itemize{
\item \code{"coding-style-notice.Rmd"}
\item \code{"installation-notice_dev-version.Rmd"}
\item \code{"installation-notice_dev-version_gitlab.Rmd"}
\item \code{"rmd-package-notice.Rmd"})`
}}
}
\value{
A \link[fs:fs_path]{path}.
}
\description{
This function gives the file path to an R Markdown snippet shipped with this package.
}
\details{
The snippets can be used anywhere R Markdown input is supported. For example, you can use them as input to knitr's
\href{https://yihui.org/knitr/options/#child-documents}{\code{child} document option}:\if{html}{\out{<div class="rmd">}}\preformatted{```\{r, child = pkgsnip::rmd_snip_path("installation-notice_dev-version_gitlab.Rmd")\}
```
}\if{html}{\out{</div>}}

Or you can use them in roxygen2's \href{https://roxygen2.r-lib.org/articles/rd.html#including-external--rmd-md-files}{\verb{@includeRmd} tag}:\if{html}{\out{<div class="r">}}\preformatted{#' @includeRmd `r pkgsnip::rmd_snip_path("installation-notice_dev-version_gitlab.Rmd")`
}\if{html}{\out{</div>}}
}
\examples{
pkgsnip::rmd_snip_path("coding-style-notice.Rmd")
}
\seealso{
R Markdown file snippet functions
\code{\link{ls_rmd_snips}()}
}
\concept{rmdsnips}
