% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{roxy_labels}
\alias{roxy_labels}
\alias{roxy_lbls}
\title{Get a table of all available roxygen2 tag labels}
\usage{
roxy_labels(
  type = c("any", "description", "param", "return", "title", NA_character_)
)

roxy_lbls(
  type = c("any", "description", "param", "return", "title", NA_character_)
)
}
\arguments{
\item{type}{The label type to return. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} listing all parameter labels included in this package, together with their \code{name} which can be provided as the \code{name}
argument of \code{\link[=param_label]{param_label()}} or \code{\link[=return_label]{return_label()}}.
}
\details{
Currently, parameter labels with the following \code{type}s and \code{name}s are available:\tabular{lll}{
   \strong{type} \tab \strong{name} \tab \strong{arguments} \cr
   \code{NA} \tab \code{data} \tab  \cr
   \code{NA} \tab \code{data_cols} \tab \verb{pal::prose_ls(cols, wrap = '}')` \cr
   \code{NA} \tab \code{date} \tab  \cr
   \code{NA} \tab \code{dates} \tab  \cr
   \code{NA} \tab \code{datetime} \tab  \cr
   \code{NA} \tab \code{datetimes} \tab  \cr
   \code{NA} \tab \code{gt_obj} \tab  \cr
   \code{NA} \tab \code{path} \tab  \cr
   \code{NA} \tab \code{paths} \tab  \cr
   \code{NA} \tab \code{plotly_obj} \tab  \cr
   \code{NA} \tab \code{r_obj} \tab  \cr
   \code{NA} \tab \code{response} \tab  \cr
   \code{NA} \tab \code{strict_list} \tab  \cr
   \code{NA} \tab \code{symbol} \tab  \cr
   \code{NA} \tab \code{tabular_data} \tab  \cr
   \code{NA} \tab \code{version_nr} \tab  \cr
   \code{NA} \tab \code{opt_global_max_cache_age} \tab \code{pkg}, \code{opt_global_max_cache_age} \cr
   \code{description} \tab \code{pkg_config} \tab \code{pkg} \cr
   \code{param} \tab \code{cli_markup_support} \tab  \cr
   \code{param} \tab \code{dyn_dots_support} \tab  \cr
   \code{param} \tab \code{tidy_select_support} \tab  \cr
   \code{param} \tab \code{eol} \tab  \cr
   \code{param} \tab \code{quiet} \tab  \cr
   \code{param} \tab \code{start_date} \tab  \cr
   \code{param} \tab \code{end_date} \tab  \cr
   \code{param} \tab \code{use_cache} \tab  \cr
   \code{param} \tab \code{max_cache_age} \tab  \cr
   \code{title} \tab \code{pkg_config} \tab \code{pkg} \cr
}
}
\seealso{
roxygen2 label functions
\code{\link{description_label}()},
\code{\link{param_label}()},
\code{\link{return_label}()},
\code{\link{roxy_label}()},
\code{\link{title_label}()}
}
\concept{roxygen2label}
