% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{roxy_lbls}
\alias{roxy_lbls}
\title{Get a table of all available roxygen2 tag labels}
\usage{
roxy_lbls(type = "default")
}
\arguments{
\item{type}{The label type to return label values for. One of \code{"default"}, \code{"author"}, \code{"concept"}, \code{"description"}, \code{"details"}, \code{"field"}, \code{"format"}, \code{"keywords"}, \code{"param"}, \code{"references"}, \code{"return"}, \code{"section"}, \code{"seealso"}, \code{"slot"}, \code{"source"} or \code{"title"}.}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df]{tibble} listing all parameter labels included in this package, together with their \code{id} which can be provided as the \code{id}
argument of \code{\link[=roxy_lbl]{roxy_lbl()}}, \code{\link[=param_lbl]{param_lbl()}}, \code{\link[=return_lbl]{return_lbl()}}, etc.
}
\details{
Currently, the following roxygen2 tag labels are available:\tabular{lll}{
   \strong{id} \tab \strong{type} \tab \strong{arguments} \cr
   \code{cli_markup_support} \tab default \tab  \cr
   \code{date} \tab default \tab  \cr
   \code{dates} \tab default \tab  \cr
   \code{datetime} \tab default \tab  \cr
   \code{datetimes} \tab default \tab  \cr
   \code{dbi_connection} \tab default \tab  \cr
   \code{df_or_tibble} \tab default \tab  \cr
   \code{dm} \tab default \tab  \cr
   \code{dyn_dots_support} \tab default \tab  \cr
   \code{end_date} \tab default \tab  \cr
   \code{eol} \tab default \tab  \cr
   \code{funky_config} \tab default \tab \code{pkg} \cr
   \code{funky_config} \tab title \tab \code{pkg} \cr
   \code{ggplot2_obj} \tab default \tab  \cr
   \code{glue_chr} \tab default \tab  \cr
   \code{gt_obj} \tab default \tab  \cr
   \code{http_method} \tab default \tab \code{one_of} \cr
   \code{max_cache_age} \tab default \tab  \cr
   \code{max_tries} \tab default \tab  \cr
   \code{num_vrsn} \tab default \tab  \cr
   \code{opt_global_max_cache_age} \tab default \tab \code{pkg}, \code{global_max_cache_age} \cr
   \code{path} \tab default \tab  \cr
   \code{paths} \tab default \tab  \cr
   \code{pkg} \tab default \tab  \cr
   \code{plotly_obj} \tab default \tab  \cr
   \code{quiet} \tab default \tab  \cr
   \code{r_obj} \tab default \tab  \cr
   \code{resp_obj} \tab default \tab  \cr
   \code{show_progress} \tab default \tab  \cr
   \code{start_date} \tab default \tab  \cr
   \code{strict_list} \tab default \tab  \cr
   \code{sym} \tab default \tab  \cr
   \code{tbl_data} \tab default \tab  \cr
   \code{tibble} \tab default \tab  \cr
   \code{tibble_cols} \tab default \tab \code{cols} \cr
   \code{tibble_custom} \tab default \tab \code{custom} \cr
   \code{tidy_select_support} \tab default \tab  \cr
   \code{use_cache} \tab default \tab  \cr
}


Note that the \code{pkg} argument defaults to the name of the package from which this function is called (\code{NULL} if not run within a package environment).
}
\examples{
pkgsnip::roxy_lbls(type = "title")
}
\seealso{
roxygen2 label functions:
\code{\link{data_roxy_lbls}},
\code{\link{data_types}},
\code{\link[=description_lbl]{description_lbl()}},
\code{\link[=param_lbl]{param_lbl()}},
\code{\link[=return_lbl]{return_lbl()}},
\code{\link[=roxy_lbl]{roxy_lbl()}},
\code{\link[=title_lbl]{title_lbl()}},
\code{\link[=type]{type()}}
}
\concept{roxylbl}
