% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{roxygen_labels}
\alias{roxygen_labels}
\title{Get a table of all roxygen labels included in this package}
\usage{
roxygen_labels(type = NULL)
}
\arguments{
\item{type}{The label type(s) to return. A character vector. Valid types include
\code{return} and \code{param})``.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
This simply returns a \link[tibble:tbl_df-class]{tibble} listing all parameter labels together with their \code{name} which can be provided as \code{\link[=param_label]{param_label()}}'s \code{name}
argument.
}
\details{
Currently, parameter labels with the following \code{types} and \code{names} are available:\tabular{ll}{
   \strong{type} \tab \strong{name} \cr
   \code{return} \tab \code{data} \cr
   \code{return} \tab \code{version_nr} \cr
   \code{return} \tab \code{path} \cr
   \code{param} \tab \code{start_date} \cr
   \code{param} \tab \code{end_date} \cr
   \code{param} \tab \code{use_cache} \cr
   \code{param} \tab \code{cache_lifespan} \cr
}
}
\seealso{
roxygen2 label functions
\code{\link{param_label}()},
\code{\link{return_label}()}
}
\concept{roxygen2label}
