% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{snippet_path}
\alias{snippet_path}
\title{Get snippet file path}
\usage{
snippet_path(snippet)
}
\arguments{
\item{snippet}{The name of a snippet. Possible values include:
\itemize{
\item \code{coding-style-notice.Rmd}
\item \code{installation-notice_dev-version.Rmd}
\item \code{installation-notice_dev-version_gitlab.Rmd}
\item \code{literate-programming-notice.Rmd})`
}}
}
\value{
A path of type \link[fs:path]{fs_path}.
}
\description{
Get snippet file path
}
\examples{
pkgsnippets::snippet_path("coding-style-notice.Rmd")
}
