% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{title_lbl}
\alias{title_lbl}
\title{Get predefined \verb{@title} label}
\usage{
title_lbl(id = roxy_lbls(type = "title")$id, as_sentence = FALSE, ...)
}
\arguments{
\item{id}{Label identifier. See \code{\link[=roxy_lbls]{roxy_lbls()}} for possible values.}

\item{as_sentence}{Whether or not to format the resulting string as a full sentence, i.e. with the first letter capitalized and a period at the end.}

\item{...}{Further named arguments used to tailor the label to your needs. Not all labels require additional arguments, see \code{\link[=roxy_lbls]{roxy_lbls()}} for an
overview. If a required argument is not explicitly provided, it is searched for in the \link[base:parent.frame]{parent frames}.}
}
\value{
A \link[glue:as_glue]{glue} character scalar.
}
\description{
Returns a predefined label intended to be used to document a function's title using \link[roxygen2:roxygen2]{roxygen2}'s
\href{https://roxygen2.r-lib.org/reference/tags-rd.html}{\verb{@title}} tag. Shorthand for \link[=roxy_lbl]{roxy_lbl(type = "title")}.
}
\details{
A label can be inserted using \href{https://roxygen2.r-lib.org/articles/reuse.html#inline-code}{inline \R code} as follows:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{#' @title `r pkgsnip::title_lbl("ID")`
}\if{html}{\out{</div>}}

Note that the above only works in \href{https://www.tidyverse.org/blog/2020/03/roxygen2-7-1-0/}{roxygen2 7.1.0+}.

Currently, the following \verb{@title} labels are available:\tabular{ll}{
   \strong{id} \tab \strong{arguments} \cr
   \code{cli_markup_support} \tab  \cr
   \code{date} \tab  \cr
   \code{dates} \tab  \cr
   \code{datetime} \tab  \cr
   \code{datetimes} \tab  \cr
   \code{dbi_connection} \tab  \cr
   \code{df_or_tibble} \tab  \cr
   \code{dm} \tab  \cr
   \code{dyn_dots_support} \tab  \cr
   \code{end_date} \tab  \cr
   \code{eol} \tab  \cr
   \code{funky_config} \tab \code{pkg} \cr
   \code{ggplot2_obj} \tab  \cr
   \code{glue_chr} \tab  \cr
   \code{gt_obj} \tab  \cr
   \code{http_method} \tab \code{one_of} \cr
   \code{max_cache_age} \tab  \cr
   \code{max_tries} \tab  \cr
   \code{num_vrsn} \tab  \cr
   \code{opt_global_max_cache_age} \tab \code{pkg}, \code{global_max_cache_age} \cr
   \code{path} \tab  \cr
   \code{paths} \tab  \cr
   \code{pkg} \tab  \cr
   \code{plotly_obj} \tab  \cr
   \code{quiet} \tab  \cr
   \code{r_obj} \tab  \cr
   \code{resp_obj} \tab  \cr
   \code{show_progress} \tab  \cr
   \code{start_date} \tab  \cr
   \code{strict_list} \tab  \cr
   \code{sym} \tab  \cr
   \code{tbl_data} \tab  \cr
   \code{tibble} \tab  \cr
   \code{tibble_cols} \tab \code{cols} \cr
   \code{tibble_custom} \tab \code{custom} \cr
   \code{tidy_select_support} \tab  \cr
   \code{use_cache} \tab  \cr
}


Note that the \code{pkg} argument defaults to the name of the package from which this function is called (\code{NULL} if not run within a package environment).
}
\examples{
# some labels take additional arguments
pkgsnip::title_lbl(id = "funky_config",
                   pkg = "foo")

# note that the `pkg` argument always defaults to the current package
# (none if run from the global environment)
pkgsnip::title_lbl(id = "funky_config")
}
\seealso{
roxygen2 label functions:
\code{\link{data_roxy_lbls}},
\code{\link{data_types}},
\code{\link[=description_lbl]{description_lbl()}},
\code{\link[=param_lbl]{param_lbl()}},
\code{\link[=return_lbl]{return_lbl()}},
\code{\link[=roxy_lbl]{roxy_lbl()}},
\code{\link[=roxy_lbls]{roxy_lbls()}},
\code{\link[=type]{type()}}
}
\concept{roxylbl}
