% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{type}
\alias{type}
\title{Get predefined type hint}
\usage{
type(id = data_types$id, length = NULL, add_cr = TRUE)
}
\arguments{
\item{id}{Type identifier. One of \code{"chr"}, \code{"date"}, \code{"df"}, \code{"fct"}, \code{"int"}, \code{"lgl"}, \code{"list"}, \code{"mat"}, \code{"num"}, \code{"pg_conn"} or \code{"raw"}}

\item{length}{Length of the \R object to be documented. Omitted if \code{NULL}.}

\item{add_cr}{Whether or not to suffix the returned string with an \href{https://rstudio.github.io/r-manuals/r-exts/Writing-R-documentation-files.html#sectioning}{Rd line break (\verb{\\cr})}.}
}
\value{
A \link[glue:as_glue]{glue} character scalar.
}
\description{
Returns a predefined \R object type hint intended to be used to document a function parameter's type.
}
\examples{
pkgsnip::type(id = "lgl",
              length = 3L)

pkgsnip::type(id = "pg_conn")
}
\seealso{
roxygen2 label functions:
\code{\link{data_roxy_lbls}},
\code{\link{data_types}},
\code{\link[=description_lbl]{description_lbl()}},
\code{\link[=param_lbl]{param_lbl()}},
\code{\link[=return_lbl]{return_lbl()}},
\code{\link[=roxy_lbl]{roxy_lbl()}},
\code{\link[=roxy_lbls]{roxy_lbls()}},
\code{\link[=title_lbl]{title_lbl()}}
}
\concept{roxylbl}
