% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{abbreviations}
\alias{abbreviations}
\title{Commonly used abbreviations in R code}
\usage{
abbreviations()
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
This simply returns a \link[tibble:tbl_df-class]{tibble} listing an opinionated set of abbreviations commonly used when writing R code. It can be used as a reference,
for example, to check availability when considering using a new abbreviation in a function or argument name.
}
