% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\docType{data}
\name{data_md_snips}
\alias{data_md_snips}
\title{Markdown snippets}
\format{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{id} and \code{value}.
}
\usage{
data_md_snips
}
\description{
A \link[tibble:tbl_df-class]{tibble} with all Markdown snippets together with their \code{id}. The latter can be fed to \code{\link[=md_snip]{md_snip()}}.
}
\details{
Currently, Markdown snippets with the following \code{id}s are available:\tabular{l}{
   \strong{id} \cr
   opt_global_max_cache_age \cr
   pkg_config \cr
   pkgdown_notice \cr
   rstudio_addin_hint \cr
   usage_notice \cr
}
}
\examples{
pkgsnip::data_md_snips
}
\seealso{
Markdown snippet functions
\code{\link{md_snip}()}
}
\concept{mdsnips}
\keyword{datasets}
