% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\docType{data}
\name{data_roxy_lbls}
\alias{data_roxy_lbls}
\title{roxygen2 tag labels}
\format{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{id}, and \code{value}.
}
\usage{
data_roxy_lbls
}
\description{
A \link[tibble:tbl_df-class]{tibble} with all Markdown snippets together with their \code{id}. The latter can be fed to \code{\link[=md_snip]{md_snip()}}.
}
\examples{
pkgsnip::data_roxy_lbls
}
\seealso{
roxygen2 label functions
\code{\link{description_lbl}()},
\code{\link{param_lbl}()},
\code{\link{return_lbl}()},
\code{\link{roxy_lbls}()},
\code{\link{roxy_lbl}()},
\code{\link{title_lbl}()}
}
\concept{roxylbl}
\keyword{datasets}
