% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\docType{data}
\name{data_types}
\alias{data_types}
\title{Type hints}
\format{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{id}, \code{value}, and \code{is_scalar}.
}
\usage{
data_types
}
\description{
A \link[tibble:tbl_df-class]{tibble} with all \R object type hints together with their \code{id}. The latter can be fed to \code{\link[=type]{type()}}.
}
\examples{
pkgsnip::data_types
}
\seealso{
roxygen2 label functions
\code{\link{data_roxy_lbls}},
\code{\link{description_lbl}()},
\code{\link{param_lbl}()},
\code{\link{return_lbl}()},
\code{\link{roxy_lbl}()},
\code{\link{roxy_lbls}()},
\code{\link{title_lbl}()},
\code{\link{type}()}
}
\concept{roxylbl}
\keyword{datasets}
