% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{description_lbl}
\alias{description_lbl}
\title{Get predefined \verb{@description} label}
\usage{
description_lbl(id = roxy_lbls(type = "description")$id, ...)
}
\arguments{
\item{id}{Label identifier. See \code{\link[=roxy_lbls]{roxy_lbls()}} for possible values.}

\item{...}{Further named arguments used to tailor the label to your needs. Not all labels require additional arguments, see \code{\link[=roxy_lbls]{roxy_lbls()}} for an overview.
If a required argument is not explicitly provided, it is searched for in the \link[=parent.frame]{parent frames}.}
}
\value{
A \link[glue:as_glue]{glue} character scalar.
}
\description{
Returns a predefined label intended to be used as a function's description using \link[roxygen2:roxygen2-package]{roxygen2}'s
\href{https://roxygen2.r-lib.org/reference/tags-rd.html}{\verb{@description}} tag.
}
\details{
A label can be inserted using \href{https://roxygen2.r-lib.org/articles/rd-formatting.html#dynamic-r-code-1}{inline \R code} as follows:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{#' @description `r pkgsnip::description_lbl("ID")`
}\if{html}{\out{</div>}}

Note that the above only works in \href{https://www.tidyverse.org/blog/2020/03/roxygen2-7-1-0/}{roxygen2 7.1.0+}.

Currently, the following \verb{@description} labels are available:\tabular{ll}{
   \strong{id} \tab \strong{arguments} \cr
   \code{cli_markup_support} \tab  \cr
   \code{date} \tab  \cr
   \code{dates} \tab  \cr
   \code{datetime} \tab  \cr
   \code{datetimes} \tab  \cr
   \code{dyn_dots_support} \tab  \cr
   \code{end_date} \tab  \cr
   \code{eol} \tab  \cr
   \code{glue_chr} \tab  \cr
   \code{gt_obj} \tab  \cr
   \code{max_cache_age} \tab  \cr
   \code{num_vrsn} \tab  \cr
   \code{opt_global_max_cache_age} \tab \code{pkg}, \code{global_max_cache_age} \cr
   \code{path} \tab  \cr
   \code{paths} \tab  \cr
   \code{pkg_config} \tab \code{pkg} \cr
   \code{plotly_obj} \tab  \cr
   \code{quiet} \tab  \cr
   \code{r_obj} \tab  \cr
   \code{resp_obj} \tab  \cr
   \code{start_date} \tab  \cr
   \code{strict_list} \tab  \cr
   \code{sym} \tab  \cr
   \code{tbl_data} \tab  \cr
   \code{tibble} \tab  \cr
   \code{tibble_cols} \tab \code{cols} \cr
   \code{tidy_select_support} \tab  \cr
   \code{use_cache} \tab  \cr
}
}
\seealso{
roxygen2 label functions
\code{\link{data_roxy_lbls}},
\code{\link{param_lbl}()},
\code{\link{return_lbl}()},
\code{\link{roxy_lbls}()},
\code{\link{roxy_lbl}()},
\code{\link{title_lbl}()}
}
\concept{roxylbl}
