% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{md_snip}
\alias{md_snip}
\title{Get predefined Markdown snippet}
\usage{
md_snip(id = data_md_snips$id, ...)
}
\arguments{
\item{id}{Snippet identifier. One of \code{"opt_global_max_cache_age"}, \code{"pkg_config"}, \code{"pkg_required"}, \code{"pkgdown_notice"}, \code{"rstudio_addin_hint"} or \code{"usage_notice"}.}

\item{...}{Further named arguments used to tailor the snippet to your needs. Not all snippets require additional arguments, see \code{\link{data_md_snips}} for an overview.
If a required argument is not explicitly provided, it is searched for in the \link[=parent.frame]{parent frames}.}
}
\value{
A character scalar.
}
\description{
Returns a predefined Markdown snippet.
}
\details{
Currently, the following Markdown snippets are available:\tabular{ll}{
   \strong{id} \tab \strong{arguments} \cr
   \code{opt_global_max_cache_age} \tab  \cr
   \code{pkg_config} \tab \code{pkg} \cr
   \code{pkg_required} \tab \code{pkg} \cr
   \code{pkgdown_notice} \tab  \cr
   \code{rstudio_addin_hint} \tab  \cr
   \code{usage_notice} \tab  \cr
}
}
\examples{
pkgsnip::md_snip(id = "rstudio_addin_hint") |> cat()

# certain snips require additional args
pkgsnip::md_snip(id = "pkg_config",
                 pkg = "foo") |>
  cat()
}
\seealso{
Markdown snippet functions
\code{\link{data_md_snips}}
}
\concept{mdsnips}
