% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-GEN.R
\name{md_snip}
\alias{md_snip}
\title{Get predefined Markdown snippet}
\usage{
md_snip(name = md_snips()$name, ...)
}
\arguments{
\item{name}{The snippet name.}

\item{...}{Further named arguments used to tailor the snippet to your needs. Not all snippets require additional arguments, see \code{\link[=md_snips]{md_snips()}} for an overview.
If a required argument is not explicitly provided, it is searched for in the \link[=parent.frame]{parent frames}.}
}
\value{
A character scalar.
}
\description{
Get predefined Markdown snippet
}
\seealso{
Markdown snippet functions
\code{\link{md_snips}()}
}
\concept{mdsnips}
