% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-GEN.R
\name{messages}
\alias{messages}
\alias{msgs}
\title{Get a table of all available \R condition messages}
\usage{
messages()

msgs()
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} listing all \R condition messages included in this package, together with their \code{name} which can be provided as
\code{\link[=message]{message()}}'s \code{name} argument.
}
\details{
Currently, \R condition messages with the following \code{names} and \code{arguments} are available:\tabular{ll}{
   \strong{name} \tab \strong{arguments} \cr
   \code{pkg_required} \tab \code{this_pkg} \cr
}
}
\seealso{
R message functions
\code{\link{message}()}
}
\concept{rmsg}
