% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{msg}
\alias{msg}
\title{Get predefined \R condition message}
\usage{
msg(name = msgs()$name, ...)
}
\arguments{
\item{name}{The message name. See \code{\link[=msgs]{msgs()}} for possible values.}

\item{...}{Further named arguments used to tailor the message to your needs. Not all messages require additional arguments, see \code{\link[=msgs]{msgs()}} for an
overview. If a required argument is not explicitly provided, it is searched for in the \link[=parent.frame]{parent frames}.}
}
\value{
A character scalar.
}
\description{
Get predefined \R condition message
}
\examples{
pkgsnip::msg(name = "pkg_required",
             this_pkg = "some_pkg")
}
\seealso{
R message functions
\code{\link{msgs}()}
}
\concept{rmsg}
