% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{msg}
\alias{msg}
\title{Get predefined R condition message}
\usage{
msg(name = messages()$name, ...)
}
\arguments{
\item{name}{The message name. See \code{\link[=messages]{messages()}} for possible values.}

\item{...}{Further named arguments used to tailor the message to your needs. Not all messages require additional arguments, see \code{\link[=messages]{messages()}} for an overview.}
}
\value{
A character scalar.
}
\description{
Get predefined R condition message
}
\examples{
pkgsnippets::msg(name = "pkg_required",
                 pkg = "some_pkg")
}
\seealso{
R message functions
\code{\link{messages}()}
}
\concept{rmsg}
