% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{param_label}
\alias{param_label}
\alias{param_lbl}
\title{Get predefined parameter label}
\usage{
param_label(name = roxy_labels(type = "param")$name, ...)

param_lbl(name = roxy_labels(type = "param")$name, ...)
}
\arguments{
\item{name}{The label name. See \code{\link[=roxy_labels]{roxy_labels()}} for possible values.}

\item{...}{Further named arguments used to tailor the label to your needs. Not all labels require additional arguments, see \code{\link[=roxy_labels]{roxy_labels()}} for an overview.
If a required argument is not explicitly provided, it is searched for in the \link[=parent.frame]{parent frames}.}
}
\value{
A character vector.
}
\description{
Returns a pre-defined label intended to be used to document function parameters using \link[roxygen2:roxygen2-package]{roxygen2}'s
\href{https://roxygen2.r-lib.org/articles/rd.html#functions}{\verb{@param}} tag.
}
\details{
A label can be inserted using \href{https://roxygen2.r-lib.org/articles/rd-formatting.html#dynamic-r-code-1}{inline \R code} as follows:\if{html}{\out{<div class="sourceCode r">}}\preformatted{#' @param start_date `r pkgsnip::param_label("start_date")`
}\if{html}{\out{</div>}}

Note that the above only works in \href{https://www.tidyverse.org/blog/2020/03/roxygen2-7-1-0/}{roxygen2 7.1.0+}.

Currently, the following parameter labels are available:\tabular{ll}{
   \strong{name} \tab \strong{arguments} \cr
   \code{data} \tab  \cr
   \code{date} \tab  \cr
   \code{dates} \tab  \cr
   \code{path} \tab  \cr
   \code{paths} \tab  \cr
   \code{r_obj} \tab  \cr
   \code{response} \tab  \cr
   \code{strict_list} \tab  \cr
   \code{tabular_data} \tab  \cr
   \code{version_nr} \tab  \cr
   \code{opt_max_cache_lifespan} \tab \code{this_pkg}, \code{max_cache_lifespan} \cr
   \code{cli_markup_support} \tab  \cr
   \code{dyn_dots_support} \tab  \cr
   \code{eol} \tab  \cr
   \code{quiet} \tab  \cr
   \code{start_date} \tab  \cr
   \code{end_date} \tab  \cr
   \code{use_cache} \tab  \cr
   \code{cache_lifespan} \tab  \cr
}
}
\seealso{
roxygen2 label functions
\code{\link{return_label}()},
\code{\link{roxy_labels}()},
\code{\link{roxy_label}()}
}
\concept{roxygen2label}
