% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-GEN.R
\name{param_label}
\alias{param_label}
\title{Get predefined parameter label}
\usage{
param_label(name = roxy_labels(type = "param")$name, ...)
}
\arguments{
\item{name}{The label name. See \code{\link[=roxy_labels]{roxy_labels()}} for possible values.}

\item{...}{Further named arguments used to tailor the label to your needs. Not all labels require additional arguments, see \code{\link[=roxy_labels]{roxy_labels()}} for an overview.}
}
\value{
A character vector.
}
\description{
These are pre-defined labels intended to be used to document function parameters using \link[roxygen2:roxygen2-package]{roxygen2}'s
\href{https://roxygen2.r-lib.org/articles/rd.html#functions}{\verb{@param}} tag.
}
\details{
The following parameter labels are available to be inserted using \href{https://roxygen2.r-lib.org/articles/rd-formatting.html#dynamic-r-code-1}{inline R code}
as follows:

\verb{#' @param data `r pkgsnip::param_label("data")` }
\verb{#' @param date `r pkgsnip::param_label("date")` }
\verb{#' @param dates `r pkgsnip::param_label("dates")` }
\verb{#' @param path `r pkgsnip::param_label("path")` }
\verb{#' @param paths `r pkgsnip::param_label("paths")` }
\verb{#' @param r_object `r pkgsnip::param_label("r_object")` }
\verb{#' @param strict_list `r pkgsnip::param_label("strict_list")` }
\verb{#' @param version_nr `r pkgsnip::param_label("version_nr")` }
\verb{#' @param opt_max_cache_lifespan `r pkgsnip::param_label("opt_max_cache_lifespan")` }
\verb{#' @param dyn_dots_support `r pkgsnip::param_label("dyn_dots_support")` }
\verb{#' @param start_date `r pkgsnip::param_label("start_date")` }
\verb{#' @param end_date `r pkgsnip::param_label("end_date")` }
\verb{#' @param use_cache `r pkgsnip::param_label("use_cache")` }
\verb{#' @param cache_lifespan `r pkgsnip::param_label("cache_lifespan")` }

Note that the above only works in \href{https://www.tidyverse.org/blog/2020/03/roxygen2-7-1-0/}{roxygen2 7.1.0+}.
}
\seealso{
roxygen2 label functions
\code{\link{return_label}()},
\code{\link{roxy_labels}()},
\code{\link{roxy_label}()}
}
\concept{roxygen2label}
