% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{return_label}
\alias{return_label}
\title{Get predefined return label}
\usage{
return_label(name = roxygen_labels(type = "return")$name)
}
\arguments{
\item{name}{The label name. See \code{\link[=roxygen_labels]{roxygen_labels()}} for possible values.}
}
\value{
A character scalar.
}
\description{
These are pre-defined labels intended to be used to document function return values using \link[roxygen2:roxygen2]{roxygen2}'s
\href{https://roxygen2.r-lib.org/articles/rd.html#functions}{\verb{@return}} tags.
}
\details{
The parameters can be inserted using \href{https://roxygen2.r-lib.org/articles/rd-formatting.html#dynamic-r-code-1}{inline R code} as follows:

\verb{#' @return `r pkgsnippets::return_label("data")` }\cr
\verb{#' @return `r pkgsnippets::return_label("path")` }\cr
}
\seealso{
roxygen2 label functions
\code{\link{md_snippets}()},
\code{\link{param_label}()},
\code{\link{roxygen_labels}()}
}
\concept{roxygen2label}
