% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip-GEN.R
\name{roxy_labels}
\alias{roxy_labels}
\title{Get a table of all available roxygen2 tag labels}
\usage{
roxy_labels(type = NULL)
}
\arguments{
\item{type}{The label type(s) to return. A character vector. Valid types include \code{"param"} and \code{"return"}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
This simply returns a \link[tibble:tbl_df-class]{tibble} listing all parameter labels included in this package, together with their \code{name} which can be provided as
the \code{name} argument of \code{\link[=param_label]{param_label()}} or \code{\link[=return_label]{return_label()}}.
}
\details{
Currently, parameter labels with the following \code{types} and \code{names} are available:\tabular{ll}{
   \strong{type} \tab \strong{name} \cr
   \code{NA} \tab \code{data} \cr
   \code{NA} \tab \code{path} \cr
   \code{NA} \tab \code{r_object} \cr
   \code{NA} \tab \code{version_nr} \cr
   \code{param} \tab \code{dyn_dots_support} \cr
   \code{param} \tab \code{start_date} \cr
   \code{param} \tab \code{end_date} \cr
   \code{param} \tab \code{use_cache} \cr
   \code{param} \tab \code{cache_lifespan} \cr
}
}
\seealso{
roxygen2 label functions
\code{\link{param_label}()},
\code{\link{return_label}()},
\code{\link{roxy_label}()}
}
\concept{roxygen2label}
