% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{roxy_lbl}
\alias{roxy_lbl}
\title{Get predefined roxygen2 tag label}
\usage{
roxy_lbl(id = roxy_lbls()$id, type = "default", as_sentence = TRUE, ...)
}
\arguments{
\item{id}{Label identifier. See \code{\link[=roxy_lbls]{roxy_lbls()}} for possible values.}

\item{type}{The label type to return the label value for. One of \code{"default"}, \code{"author"}, \code{"concept"}, \code{"description"}, \code{"details"}, \code{"field"}, \code{"format"}, \code{"keywords"}, \code{"param"}, \code{"references"}, \code{"return"}, \code{"section"}, \code{"seealso"}, \code{"slot"}, \code{"source"} or \code{"title"}.}

\item{as_sentence}{Whether or not to format the resulting string as a full sentence, i.e. with the first letter capitalized and a period at the end.}

\item{...}{Further named arguments used to tailor the label to your needs. Not all labels require additional arguments, see \code{\link[=roxy_lbls]{roxy_lbls()}} for an
overview. If a required argument is not explicitly provided, it is searched for in the \link[=parent.frame]{parent frames}.}
}
\value{
A \link[glue:as_glue]{glue} character scalar.
}
\description{
Returns a predefined label intended to be used to document functions using \link[roxygen2:roxygen2-package]{roxygen2}
\href{https://roxygen2.r-lib.org/articles/rd.html#functions}{tags}.
}
\details{
The labels can be inserted using \href{https://roxygen2.r-lib.org/articles/reuse.html#inline-code}{inline \R code} as follows:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{#' @param start_date `r pkgsnip::roxy_lbl("start_date", type = "param")`
#' @return `r pkgsnip::roxy_lbl("tibble", type = "return")`
}\if{html}{\out{</div>}}

Note that the above only works in \href{https://www.tidyverse.org/blog/2020/03/roxygen2-7-1-0/}{roxygen2 7.1.0+}.
}
\seealso{
roxygen2 label functions
\code{\link{data_roxy_lbls}},
\code{\link{data_types}},
\code{\link{description_lbl}()},
\code{\link{param_lbl}()},
\code{\link{return_lbl}()},
\code{\link{roxy_lbls}()},
\code{\link{title_lbl}()},
\code{\link{type}()}
}
\concept{roxylbl}
\keyword{internal}
