% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{roxygen_labels}
\alias{roxygen_labels}
\title{Get a table of all roxygen labels included in this package}
\usage{
roxygen_labels(type = NULL)
}
\arguments{
\item{type}{The label type(s) to return. A character vector. Valid types include
\code{return} and \code{param})``.}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
This simply returns a \link[tibble:tbl_df]{tibble} listing all parameter labels together with their \code{name} which can be provided as \code{\link[=param_label]{param_label()}}'s \code{name}
argument.
}
\details{
Currently, parameter labels with the following \code{types} and \code{names} are available:

\link{1} "\n\n| \strong{type} | \strong{name}         |\n|:---------|:-----------------|\n| \code{return} | \code{data}           |\n| \code{return} | \code{path}           |\n| \code{param}  | \code{start_date}     |\n| \code{param}  | \code{end_date}       |\n| \code{param}  | \code{use_cache}      |\n| \code{param}  | \code{cache_lifespan} |\n\n"
}
\seealso{
roxygen2 label functions
\code{\link{md_snippets}()},
\code{\link{param_label}()},
\code{\link{return_label}()}
}
\concept{roxygen2label}
