% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnip.gen.R
\name{snip_path}
\alias{snip_path}
\title{Get R Markdown snippet file path}
\usage{
snip_path(id = ls_file_snips()$id)
}
\arguments{
\item{id}{Snippet identifier. One of \code{"coding-style-notice.Rmd"}, \code{"common.Rmd"}, \code{"installation-notice_dev-version.Rmd"}, \code{"installation-notice_dev-version_gitlab.Rmd"} or \code{"rmd-package-notice.Rmd"}}
}
\value{
A \link[fs:fs_path]{path}.
}
\description{
Returns the file path to an R Markdown snippet shipped with this package.
}
\details{
The snippets can be used anywhere R Markdown input is supported. For example, you can use them as input to knitr's
\href{https://yihui.org/knitr/options/#child-documents}{\code{child} document option}:

\if{html}{\out{<div class="sourceCode rmd">}}\preformatted{```\{r, child = pkgsnip::snip_path("installation-notice_dev-version_gitlab.Rmd")\}
```
}\if{html}{\out{</div>}}

Or you can use them in roxygen2's \href{https://roxygen2.r-lib.org/articles/rd.html#including-external--rmd-md-files}{\verb{@includeRmd} tag}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{#' @includeRmd `r pkgsnip::snip_path("installation-notice_dev-version_gitlab.Rmd")`
}\if{html}{\out{</div>}}
}
\examples{
pkgsnip::snip_path("coding-style-notice.Rmd")
}
\seealso{
R Markdown file snippet functions
\code{\link{ls_file_snips}()}
}
\concept{rmdsnips}
