% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsnippets-GEN.R
\name{snippet_path}
\alias{snippet_path}
\title{Get snippet file path}
\usage{
snippet_path(name = ls_snippets()$name)
}
\arguments{
\item{name}{The name of a snippet. Possible values include:
\itemize{
\item \code{"coding-style-notice.Rmd"}
\item \code{"installation-notice_dev-version.Rmd"}
\item \code{"installation-notice_dev-version_gitlab.Rmd"}
\item \code{"rmd-package-notice.Rmd"})`
}}
}
\value{
A \link[fs:fs_path]{path}.
}
\description{
This function gives the file path to an (R) Markdown snippet shipped with this package.
}
\details{
The snippets can be used anywhere (R) Markdown input is supported. For example, you can use them as input to knitr's
\href{https://yihui.org/knitr/options/#child-documents}{\code{child} document option}:\if{html}{\out{<div class="rmd">}}\preformatted{```\{r, child = pkgsnippets::snippet_path("installation-notice_dev-version_gitlab.Rmd")\}
```
}\if{html}{\out{</div>}}

Or you can use them in roxygen2's \href{https://roxygen2.r-lib.org/articles/rd.html#including-external--rmd-md-files}{\verb{@includeRmd} tag}:\if{html}{\out{<div class="r">}}\preformatted{#' @includeRmd `r pkgsnippets::snippet_path("installation-notice_dev-version_gitlab.Rmd")`
}\if{html}{\out{</div>}}
}
\examples{
pkgsnippets::snippet_path("coding-style-notice.Rmd")
}
